test: move narrowing/Optional tests into their own module
The flow-narrowing + Optional/Undef tests (flow guards, places, negatives,
defined/blessed strip, the lattice join/ternary/Maybe) move out of the
general builder_tests.rs bulk into src/builder/narrowing_tests.rs, next to
the narrowing.rs they exercise. A child of the `tests` module so it shares
build_fa; the invocant_class_of helper moves with it. Pure relocation —
36 tests now under tests::narrowing::*, 983 total green, no warnings.
